extract sva_sequence_matcht
This extracts the class that represents potential SVA sequence matches, and
the function that generates these from SVA sequence expressions into a
separate file.
