+++ src/temporal-logic/sva_sequence_match.cpp
@@ -1,0 +1,122 @@
+/*******************************************************************\
+
+Module: SVA Sequence Matches
+
+Author: Daniel Kroening, dkr@amazon.com
+
+\*******************************************************************/
+
+#include "sva_sequence_match.h"
+
+#include <util/arith_tools.h>
+#include <util/std_expr.h>
+
+#include <verilog/sva_expr.h>
+
+sva_sequence_matcht sva_sequence_matcht::true_match(const mp_integer &n)
+{
+  sva_sequence_matcht result;
+  for(mp_integer i; i < n; ++i)
+    result.cond_vector.push_back(true_exprt{});
+  return result;
+}
+
+// nonoverlapping concatenation
+sva_sequence_matcht concat(sva_sequence_matcht a, const sva_sequence_matcht &b)
+{
+  a.cond_vector.insert(
+    a.cond_vector.end(), b.cond_vector.begin(), b.cond_vector.end());
+  return a;
+}
+
+// overlapping concatenation
+sva_sequence_matcht
+overlapping_concat(sva_sequence_matcht a, sva_sequence_matcht b)
+{
+  PRECONDITION(!a.empty_match());
+  PRECONDITION(!b.empty_match());
+  auto a_last = a.cond_vector.back();
+  a.cond_vector.pop_back();
+  b.cond_vector.front() = conjunction({a_last, b.cond_vector.front()});
+  return concat(std::move(a), b);
+}
+
+std::vector<sva_sequence_matcht> LTL_sequence_matches(const exprt &sequence)
+{
+  if(sequence.id() == ID_sva_boolean)
+  {
+    // atomic proposition
+    return {sva_sequence_matcht{to_sva_boolean_expr(sequence).op()}};
+  }
+  else if(sequence.id() == ID_sva_sequence_concatenation)
+  {
+    auto &concatenation = to_sva_sequence_concatenation_expr(sequence);
+    auto matches_lhs = LTL_sequence_matches(concatenation.lhs());
+    auto matches_rhs = LTL_sequence_matches(concatenation.rhs());
+
+    if(matches_lhs.empty() || matches_rhs.empty())
+      return {};
+
+    std::vector<sva_sequence_matcht> result;
+
+    // cross product
+    for(auto &match_lhs : matches_lhs)
+      for(auto &match_rhs : matches_rhs)
+      {
+        // Sequence concatenation is overlapping
+        auto new_match = overlapping_concat(match_lhs, match_rhs);
+        CHECK_RETURN(
+          new_match.length() == match_lhs.length() + match_rhs.length() - 1);
+        result.push_back(std::move(new_match));
+      }
+    return result;
+  }
+  else if(sequence.id() == ID_sva_cycle_delay)
+  {
+    auto &delay = to_sva_cycle_delay_expr(sequence);
+    auto matches = LTL_sequence_matches(delay.op());
+    auto from_int = numeric_cast_v<mp_integer>(delay.from());
+
+    if(matches.empty())
+      return {};
+
+    if(delay.to().is_nil())
+    {
+      // delay as instructed
+      auto delay_sequence = sva_sequence_matcht::true_match(from_int);
+
+      for(auto &match : matches)
+        match = concat(delay_sequence, match);
+
+      return matches;
+    }
+    else if(delay.to().id() == ID_infinity)
+    {
+      return {}; // can't encode
+    }
+    else if(delay.to().is_constant())
+    {
+      auto to_int = numeric_cast_v<mp_integer>(to_constant_expr(delay.to()));
+      std::vector<sva_sequence_matcht> new_matches;
+
+      for(mp_integer i = from_int; i <= to_int; ++i)
+      {
+        // delay as instructed
+        auto delay_sequence = sva_sequence_matcht::true_match(i);
+
+        for(const auto &match : matches)
+        {
+          new_matches.push_back(concat(delay_sequence, match));
+        }
+      }
+
+      return new_matches;
+    }
+    else
+      return {};
+  }
+  else
+  {
+    return {}; // unsupported
+  }
+}
